All: Added lightstyle support across modules. This allows the game module to parse the lightmap styles from the worldspawn entity, which the game can then communicate to the cgame module through the use of the added CS_LIGHT_STYLES config strings.
New functionality is added to the cgame module as well, updating the style
colors in the active renderer using the config strings provided by the game
module.

The RE_GetLightStyle and RE_SetLightStyle renderer entry points are added to
both the OpenGL1 and OpenGL2 renderers.

Please note that the lightstyle support in the cgame and game modules are
temporary additions. They only exist so functionality in the OpenGL1 and OpenGL2
renderers can be tested and used. Once iosof2mp advances to a later stadium the
original SoF2MP SDK (containing the cgame, game and ui modules) will be added to
iosof2mp instead.