OpenGL1: Changed RB_ComputeFinalVertexColor function. Changed the function return type from ulong to void. ulong isn't supported anymore when compiling iosof2mp with MinGW, so the ulong type has been replaced by uint32_t.
To maintain consistency between the OpenGL1 and OpenGL2 renderers, the function
now takes a uint8_t pointer type as input (colors are 16 bits in OpenGL2). The
output is now also set in the function instead of casting the bytes in place.
That means an additional output parameter has been added.

Also new is the forceValidStyle parameter. If the style color is unused, the
color style is not applied and the input is used instead. This is the exact same
behavior as vanilla SoF2MP.