OpenGL2: Changed overbright lighting. This change limits overbrighting by forcing the "r_overBrightBits" CVAR between 0 (default value) and 1. The "r_mapOverBrightBits" CVAR is now also forced between 0 and 1 (default value).
Additionally, the maximum allowed overbright bits set in R_SetColorMappings is
now 1. The R_ColorShiftLightingBytes function is also modified to only alter the
color data under certain conditions, otherwise the function will copy unaltered
input color data to the output array.

This results in almost the exact same behavior as vanilla SoF2MP, with the
exception that the OpenGL2 renderer doesn't allow more overbright bits than map
overbright bits in R_SetColorMappings. This results in different behavior
between the OpenGL2 renderer and the OpenGL1 renderer/vanilla SoF2MP when
r_overBrightBits is set to 1 and r_mapOverBrightBits is set to 0.